# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexusProxy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  clerkId: string;
  email: string;
  role: string;
  isBanned: boolean;
  createdAt: string;
}

export interface Plan {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  planType: string;
  priceUsd: number;
  bandwidthGb: number;
  proxyCount: number;
  durationDays: number;
  proxyTypes: string[];
  features: string[];
  isActive: boolean;
}

export interface Payment {
  id: string;
  userId: string;
  planId: string;
  amountUsd: number;
  currency: string;
  /** @nullable */
  cryptoAmount?: string | null;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  txHash?: string | null;
  status: string;
  /** @nullable */
  confirmedAt?: string | null;
  createdAt: string;
}

export interface CreatePaymentBody {
  planId: string;
  currency: string;
}

export interface SubmitHashBody {
  txHash: string;
}

export interface Subscription {
  id: string;
  userId: string;
  planId: string;
  status: string;
  bandwidthGbTotal: number;
  bandwidthUsedMb: number;
  /** @nullable */
  startsAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface UserProxy {
  id: string;
  proxyId: string;
  ip: string;
  port: number;
  username: string;
  password: string;
  proxyType: string;
  /** @nullable */
  country?: string | null;
  isActive: boolean;
  assignedAt: string;
}

export interface ActiveSubscription {
  subscription: Subscription;
  plan: Plan;
  proxies: UserProxy[];
  /** @nullable */
  remainingHours?: number | null;
  bandwidthRemainingMb: number;
}

export interface UsageStats {
  totalBandwidthUsedMb: number;
  activeProxies: number;
  totalConnections: number;
  /** @nullable */
  lastActivity?: string | null;
}

export interface AdminUser {
  id: string;
  clerkId: string;
  email: string;
  role: string;
  isBanned: boolean;
  createdAt: string;
  /** @nullable */
  activeSubscription?: string | null;
}

export interface BanUserBody {
  isBanned: boolean;
}

export interface AdminPayment {
  id: string;
  userId: string;
  /** @nullable */
  userEmail?: string | null;
  planId: string;
  /** @nullable */
  planName?: string | null;
  amountUsd: number;
  currency: string;
  /** @nullable */
  txHash?: string | null;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}

export interface ConfirmPaymentBody {
  adminNote?: string;
}

export interface AdminProxy {
  id: string;
  ip: string;
  port: number;
  username: string;
  proxyType: string;
  /** @nullable */
  country?: string | null;
  isActive: boolean;
  isAssigned: boolean;
  createdAt: string;
}

export interface AddProxyBody {
  ip: string;
  port: number;
  username: string;
  password: string;
  proxyType: string;
  country?: string;
}

export interface BulkAddProxiesBody {
  /** Newline-separated proxies in IP:PORT:USER:PASS format */
  proxyList: string;
  proxyType: string;
}

export interface BulkAddResult {
  added: number;
  skipped: number;
  errors: string[];
}

export interface AdminStats {
  totalUsers: number;
  activeSubscriptions: number;
  pendingPayments: number;
  totalProxies: number;
  assignedProxies: number;
  totalRevenueCents: number;
}

export interface CreatePlanBody {
  id: string;
  name: string;
  description?: string;
  planType: string;
  priceUsd: number;
  bandwidthGb: number;
  proxyCount: number;
  durationDays: number;
  proxyTypes?: string[];
  features?: string[];
}